Protocols: Create posix.Symlink to create links on download rucio #5319
Subclass of the Default posix implemenation to allow a symlink to be made to the remote location,
instead of copying the remote file locally.

Only the get() method is overridden to provide this functionality. Other methods use the default implementation,
which involve using the remote file directly.

Addtional tests for this implementation are included into the posix test code.